Prereq: move xz to aws, standardized prereq install scripts

Closes #1371

See merge request opencpi/opencpi!363

+++ build/prerequisites/gpsd/install-gpsd.sh
@@ -1,4 +1,4 @@
-#!/bin/bash --noprofile
+#!/bin/bash
 # This file is protected by Copyright. Please refer to the COPYRIGHT file
 # distributed with this source distribution.
 #
@@ -20,40 +20,47 @@
 # This server is unavailable:       https://ftp.gnu.org/gnu/gmp
 # Since we don't look at multiple URLs/mirrors (yet)
 # The one below is one of the advertised mirrors
-me=gpsd
 
-[ -z "$OCPI_CDK_DIR" ] && echo Environment variable OCPI_CDK_DIR not set && exit 1
+[ -z "$OCPI_CDK_DIR" ] && echo 'Environment variable OCPI_CDK_DIR not set' && exit 1
 
-version=release-3.18 # NOTE: when updating gpsd version, "${OcpiCrossCompile}ar"
-                     # command below may be affected
-# Download gpsd source using git
-source $OCPI_CDK_DIR/scripts/setup-prerequisite.sh \
-       "$1" \
-       $me \
-       "GPS (Global Positioning System) daemon" \
-       https://gitlab.com/gpsd/gpsd.git \
-       $version \
-       $me \
-       1
+target_platform="$1"
+name=gpsd
+version=3.18 # NOTE: when updating gpsd version, "${OcpiCrossCompile}ar"
+             # command below may be affected
+#version=3.21  # latest as of 09/02/2020
+pkg_name="$name-$version"
+description='GPS (Global Positioning System) daemon'
+dl_url="http://download-mirror.savannah.gnu.org/releases/$name/${pkg_name}.tar.gz"
+extracted_dir="$pkg_name"
+cross_build=1
 
-# gpsd does not support separate build directories yet, even though scons does, so we have to copy everything,
-# per-platform, to the build directory for the platform
-here=$(basename $PWD)
-echo Making a copy of gpsd for platform $platform because gpsd does not yet support separate build directories.
-(cd ..; for i in *; do [[ $i != ocpi-build-* ]] && cp -R -p $i $here; done)
+# Download and extract source
+source "$OCPI_CDK_DIR/scripts/setup-prerequisite.sh" \
+       "$target_platform" \
+       "$name" \
+       "$description" \
+       "${dl_url%/*}" \
+       "${dl_url##*/}" \
+       "$extracted_dir" \
+       "$cross_build"
 
-pwd
-echo $OcpiThisPrerequisiteDir
-ls -alF .
+# GPSd does not support separate build directories yet, even though scons does,
+# so we have to copy everything, per-platform, to the build directory for the
+# platform
+here="$(basename "$PWD")"
+echo "Making a copy of gpsd for platform $platform because gpsd does not yet \
+support separate build directories."
+(cd ..; for i in *; do [[ "$i" != ocpi-build-* ]] && cp -Rp "$i" "$here"; done)
+
 patchfile=compiler.patch
-patch -p0 < $OcpiThisPrerequisiteDir/compiler.patch || {
-  echo "*******************************************************" >&2
-  echo "ERROR: patch applied by compiler.patch failed!!" >&2
-  echo "*******************************************************" >&2
+patch -p0 < "$OcpiThisPrerequisiteDir/$patchfile" || {
+  echo '*******************************************************' >&2
+  echo "ERROR: patch applied by $patchfile failed!!" >&2
+  echo '*******************************************************' >&2
   exit 1
 }
 
-# framework will error out w/ "recompile with -fPIC" if these aren't included
+# Framework will error out w/ "recompile with -fPIC" if these aren't included
 export CFLAGS="-fPIC"
 export CXXFLAGS="-fPIC"
 
@@ -62,36 +69,37 @@ export CXXFLAGS="-fPIC"
 #              linkable.
 # nostrip=True is necessary to avoid frameworker error during build: "error
 #              adding symbols: Archive has no index; run ranlib to add one".
-echo OcpiCrossCompile=$OcpiCrossCompile
-echo OcpiCrossHost=$OcpiCrossHost
+echo "OcpiCrossCompile=$OcpiCrossCompile"
+echo "OcpiCrossHost=$OcpiCrossHost"
 
 # Set sysroot if we are cross building
 sysroot=
-[ -n "$OcpiCrossHost" ] && sysroot=$(echo $OcpiCrossCompile | sed "s|/bin[^/]*/$OcpiCrossHost-||")/$OcpiCrossHost/libc
+[ -n "$OcpiCrossHost" ] && sysroot="$(echo "$OcpiCrossCompile" | sed "s|/bin[^/]*/$OcpiCrossHost-||")/$OcpiCrossHost/libc"
 
-#
 # Figure out how a potentially python3-compatible version of SCons
 # is invoked on $OCPI_TOOL_PLATFORM.  As of 12 May 2020, there are
 # two known possibilities.  If and/or when it becomes necessary to
 # check for others, the preferred order is from "least likely" to
 # "most likely", i.e., the sensible default of "scons" should be
 # the last possibility checked.
-#
-SCons=`command -v scons-3 || command -v scons`
+SCONS="$(command -v scons-3 || command -v scons)"
+if [ -z "$SCONS" ]; then
+  echo "Error: cannot find scons or scons-3 which are required to build $name"
+  exit 1
+fi
 
-#echo "DEBUG: SCons is \"$SCons\""
+"$SCONS" prefix="$OcpiInstallExecDir" target="$OcpiCrossHost" \
+         sysroot="$sysroot" libgpsmm=True ncurses=False qt=False python=False \
+         usb=False bluez=False ntp=False manbuild=False shared=False \
+         nostrip=True debug=True dbus_export=False
+"$SCONS" install
 
-$SCons prefix=$OcpiInstallExecDir target=$OcpiCrossHost sysroot=$sysroot \
-      libgpsmm=True ncurses=False qt=False python=False usb=False \
-      bluez=False ntp=False manbuild=False shared=False nostrip=True \
-      debug=True dbus_export=False
-$SCons install
 # Because OpenCPI prerequisites must be statically compiled into a single
 # archive, and the scons build produces both libgps.a and libgpsd.a, a single
 # archive is manually constructed. The arguments to pass to ar were determined
 # by: 1) run scons build, 2) observe two ar commands sent to stdout, 3) combine
 # ar commands into one, 4) replace all *.a with a single libtmp.a
-${OcpiCrossCompile}ar rc libtmp.a ais_json.o bits.o gpsdclient.o \
+"${OcpiCrossCompile}ar" rc libtmp.a ais_json.o bits.o gpsdclient.o \
     gps_maskdump.o gpsutils.o \
     hex.o json.o libgps_core.o libgps_dbus.o libgps_json.o \
     libgps_shm.o libgps_sock.o netlib.o os_compat.o \
@@ -108,14 +116,15 @@ ${OcpiCrossCompile}ar rc libtmp.a ais_json.o bits.o gpsdclient.o \
     net_gnss_dispatch.o net_ntrip.o ntpshmread.o ntpshmwrite.o packet.o \
     ppsthread.o pseudoais.o pseudonmea.o serial.o subframe.o timebase.o \
     timespec_str.o
-${OcpiCrossCompile}ranlib libtmp.a
-rm -rf $OcpiInstallExecDir/lib/*
-cp libtmp.a $OcpiInstallExecDir/lib/libgpsd.a
+"${OcpiCrossCompile}ranlib" libtmp.a
 
-echo "copying lower level headers which are necessary for use of libgpsd directly..."
+rm -rf "${OcpiInstallExecDir:?}/lib/*"
+cp libtmp.a "$OcpiInstallExecDir/lib/libgpsd.a"
+
+echo 'copying lower level headers which are necessary for use of libgpsd \
+directly...'
 headers=(gpsd.h compiler.h gpsd_config.h gps.h ppsthread.h os_compat.h)
 for header in "${headers[@]}"; do
   echo "...copying $header to $OcpiInstallExecDir/include/"
-  cp $header $OcpiInstallExecDir/include/
+  cp "$header" "$OcpiInstallExecDir/include/"
 done
-
